backend/app/services/gift_card_service.py
"""
Gift Card Service - Handles integration with third-party gift card providers

This service provides a unified interface for purchasing and delivering gift cards
from various providers like Amazon, Netflix, Spotify, etc.

Integration Options:
1. Tango Card API (https://www.tangocard.com/) - Aggregator for multiple brands
2. Giftbit API (https://www.giftbit.com/) - Gift card marketplace
3. Blackhawk Network - Enterprise gift card solutions
4. Direct API integrations (if available from providers)

For now, this is a placeholder structure that can be implemented when you:
1. Sign up for a gift card API service
2. Get API credentials
3. Implement the actual API calls
"""

from typing import Optional, Dict, Any
from decimal import Decimal
from enum import Enum
import logging

logger = logging.getLogger(__name__)


class GiftCardProvider(str, Enum):
    """Supported gift card providers"""
    AMAZON = "amazon"
    NETFLIX = "netflix"
    SPOTIFY = "spotify"
    APPLE = "apple"
    GOOGLE_PLAY = "google_play"
    UBER_EATS = "uber_eats"
    STARBUCKS = "starbucks"
    AIRBNB = "airbnb"
    STEAM = "steam"
    DISNEY_PLUS = "disney_plus"
    SEPHORA = "sephora"
    NIKE = "nike"
    UBER = "uber"
    DOORDASH = "doordash"
    MASTERCLASS = "masterclass"


class GiftCardService:
    """
    Service for purchasing and managing gift cards from third-party providers
    
    This is a placeholder implementation. To make it work:
    
    1. Choose a gift card API provider (recommended: Tango Card or Giftbit)
    2. Sign up and get API credentials
    3. Install their SDK or use their REST API
    4. Implement the purchase_gift_card method
    5. Add error handling and webhook support for delivery status
    """
    
    def __init__(self):
        # TODO: Initialize API client with credentials from environment variables
        # Example:
        # self.tango_client = TangoCardClient(
        #     api_key=os.getenv("TANGO_CARD_API_KEY"),
        #     environment=os.getenv("TANGO_CARD_ENV", "sandbox")
        # )
        pass
    
    async def purchase_gift_card(
        self,
        provider: GiftCardProvider,
        amount: Decimal,
        currency: str = "USD",
        recipient_email: Optional[str] = None,
        recipient_name: Optional[str] = None,
        message: Optional[str] = None
    ) -> Dict[str, Any]:
        """
        Purchase a gift card from a provider
        
        Args:
            provider: The gift card provider (Amazon, Netflix, etc.)
            amount: The gift card amount
            currency: Currency code (default: USD)
            recipient_email: Email to send the gift card to
            recipient_name: Name of the recipient
            message: Optional message to include
            
        Returns:
            Dict containing:
                - success: bool
                - gift_card_code: str (the actual code)
                - gift_card_url: str (link to redeem)
                - provider_reference: str (provider's transaction ID)
                - error: Optional[str]
        """
        
        # TODO: Implement actual API call
        # Example with Tango Card:
        # try:
        #     response = await self.tango_client.create_order(
        #         account_identifier=provider.value,
        #         amount=amount,
        #         recipient_email=recipient_email,
        #         message=message
        #     )
        #     return {
        #         "success": True,
        #         "gift_card_code": response.gift_card_code,
        #         "gift_card_url": response.gift_card_url,
        #         "provider_reference": response.order_id
        #     }
        # except Exception as e:
        #     logger.error(f"Failed to purchase gift card: {e}")
        #     return {
        #         "success": False,
        #         "error": str(e)
        #     }
        
        # Placeholder response
        logger.warning(f"Gift card purchase not implemented. Provider: {provider}, Amount: {amount}")
        return {
            "success": False,
            "error": "Gift card service not yet configured. Please integrate with a gift card API provider."
        }
    
    async def check_gift_card_status(
        self,
        provider_reference: str
    ) -> Dict[str, Any]:
        """
        Check the status of a gift card order
        
        Args:
            provider_reference: The provider's transaction/order ID
            
        Returns:
            Dict with status information
        """
        # TODO: Implement status check
        return {
            "status": "unknown",
            "error": "Status check not implemented"
        }
    
    def get_supported_providers(self) -> list[GiftCardProvider]:
        """Get list of supported gift card providers"""
        return list(GiftCardProvider)
    
    def is_provider_available(self, provider: GiftCardProvider) -> bool:
        """Check if a provider is available for purchase"""
        # TODO: Check provider availability from API
        return True


# Example integration guide for Tango Card:
"""
To integrate with Tango Card API:

1. Sign up at https://www.tangocard.com/
2. Get API credentials (API Key, Platform Name, Platform Key)
3. Install SDK: pip install tangocard-sdk
4. Update this service:

from tangocard import TangoCardClient

class GiftCardService:
    def __init__(self):
        self.client = TangoCardClient(
            api_key=os.getenv("TANGO_CARD_API_KEY"),
            platform_name=os.getenv("TANGO_CARD_PLATFORM_NAME"),
            platform_key=os.getenv("TANGO_CARD_PLATFORM_KEY"),
            environment=os.getenv("TANGO_CARD_ENV", "sandbox")
        )
    
    async def purchase_gift_card(self, provider, amount, ...):
        # Map provider to Tango Card catalog item
        catalog_item = self._map_provider_to_catalog(provider)
        
        # Create order
        order = await self.client.create_order(
            account_identifier=catalog_item,
            amount=int(amount * 100),  # Convert to cents
            recipient_email=recipient_email,
            message=message
        )
        
        return {
            "success": True,
            "gift_card_code": order.gift_card_code,
            "gift_card_url": order.gift_card_url,
            "provider_reference": order.order_id
        }
"""

