backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0
firebase-admin==6.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
stripe==8.2.0
paypalrestsdk==1.13.1
httpx==0.27.0
pillow==10.2.0
pytest==8.0.1
pytest-asyncio==0.23.5
google-generativeai==0.3.2

